# docs/swagger.yaml
basePath: /
definitions:
...
  handlers.ErrorResponse:
    properties:
      error:
        example: Invalid input
        type: string
    type: object
...
  handlers.OAuthURLResponse:
    properties:
      auth_url:
        example: https://accounts.google.com/oauth/authorize?...
        type: string
      state:
        example: randomStateString
        type: string
    type: object
  handlers.ProjectInvitationResponse:
    properties:
      created_at:
        example: "2023-01-01T00:00:00Z"
        type: string
      email:
        example: invitee@example.com
        type: string
      id:
        example: 1
        type: integer
      invitee:
        $ref: '#/definitions/handlers.UserResponse'
      invitee_id:
        example: 2
        type: integer
      inviter:
        $ref: '#/definitions/handlers.UserResponse'
      inviter_id:
        example: 1
        type: integer
      project:
        $ref: '#/definitions/handlers.ProjectResponse'
      project_id:
        example: 1
        type: integer
      role:
        example: collaborator
        type: string
      status:
        example: pending
        type: string
      updated_at:
        example: "2023-01-01T00:00:00Z"
        type: string
    type: object
  handlers.ProjectResponse:
    properties:
      created_at:
        example: "2023-01-01T00:00:00Z"
        type: string
      description:
        example: Project description
        type: string
      id:
        example: 1
        type: integer
      name:
        example: My Project
        type: string
      owner:
        $ref: '#/definitions/handlers.UserResponse'
      owner_id:
        example: 1
        type: integer
      updated_at:
        example: "2023-01-01T00:00:00Z"
        type: string
    type: object
...
  handlers.UserResponse:
    properties:
      created_at:
        example: "2023-01-01T00:00:00Z"
        type: string
      email:
        example: user@example.com
        type: string
      email_verified:
        example: true
        type: boolean
      id:
        example: 1
        type: integer
      is_active:
        example: true
        type: boolean
      last_login_at:
        example: "2023-01-01T00:00:00Z"
        type: string
      name:
        example: John Doe
        type: string
      two_factor_enabled:
        example: false
        type: boolean
      updated_at:
        example: "2023-01-01T00:00:00Z"
        type: string
    type: object
...
  services.InviteUserInput:
    properties:
      message:
        description: Optional invitation message
        type: string
      project_id:
        description: Project to invite user to
        type: integer
      user_email:
        description: Email of user to invite
        type: string
    required:
    - project_id
    - user_email
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: A comprehensive task management API with real-time collaboration features
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Real Task API
  version: "1.0"
paths:
...
  /api/auth/google/url:
    get:
      consumes:
      - application/json
      description: Get Google OAuth authorization URL for login
      produces:
      - application/json
      responses:
        "200":
          description: OAuth URL generated successfully
          schema:
            $ref: '#/definitions/handlers.OAuthURLResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Get Google OAuth URL
      tags:
      - OAuth
...
  /api/projects/{project_id}/invite:
    post:
      consumes:
      - application/json
      description: Invite a user to collaborate on a project (project owner only)
      parameters:
      - description: Project ID
        in: path
        name: project_id
        required: true
        type: integer
      - description: Invitation details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/services.InviteUserInput'
      produces:
      - application/json
      responses:
        "201":
          description: Invitation sent successfully
          schema:
            $ref: '#/definitions/handlers.ProjectInvitationResponse'
        "400":
          description: Invalid input
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: User not authenticated
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "403":
          description: Access denied
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Invite user to project
      tags:
      - Collaboration
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"